requirements.txt
numpy
ScientificPython
matplotlib
shapely
GDAL
Pillow
